[client] ds: avoid util_cursorToInt when warping pointer
Using util_cursorToInt messes with the error tracking for normal movements,
and is not necessary since we are computing an absolute position on the
client window.

Instead, we should pass doubles directly to display servers and let them
decide how to best handle them. For example, XIWarpPointer accepts doubles
directly.

+++ client/displayservers/Wayland/input.c
@@ -459,10 +459,10 @@ void waylandRealignPointer(void)
     app_resyncMouseBasic();
 }
 
-void waylandGuestPointerUpdated(double x, double y, int localX, int localY)
+void waylandGuestPointerUpdated(double x, double y, double localX, double localY)
 {
   if (!wlWm.warpSupport || !wlWm.pointerInSurface || wlWm.lockedPointer)
     return;
 
-  waylandWarpPointer(localX, localY, false);
+  waylandWarpPointer((int) localX, (int) localY, false);
 }
